+++ codigo/src/util/Point.java
@@ -2,17 +2,17 @@ package util;
 
 public class Point {
 
-	private int x, y;
+	private float x, y;
 	
-	public Point(int x, int y){
+	public Point(float x, float y){
 		this.x=x;
 		this.y=y;
 	}
 	
-	public int getX(){
+	public float getX(){
 		return x;
 	}
-	public int getY(){
+	public float getY(){
 		return y;
 	}
 }
